requirements.txt
algoliasearch>=2.0,<3.0
diagrams
gitpython
jupyterlab
matplotlib
networkx
pandas
pydantic[dotenv]
pyvis
tqdm
